+++ sduoj-submit/src/main/java/cn/edu/sdu/qd/oj/submit/controller/SubmitController.java
@@ -49,7 +49,7 @@ public class SubmitController {
     public Long createSubmission(@RequestBody Map json,
                                 HttpServletRequest request) {
         int problemId = (int) json.get("problemId");
-        Byte languageId = (byte) json.get("languageId");
+        int languageId = (int) json.get("languageId");
         String code = (String) json.get("code");
         String ipv4 = request.getHeader("X-FORWARDED-FOR");
         if (ipv4 == null) {
